[SB-37] feat: App, AppState 최초 생성
