+++ feature/main/src/main/java/co/kr/hoyaho/main/ui/SubwayApp.kt
@@ -1,0 +1,30 @@
+package co.kr.hoyaho.main.ui
+
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+@Suppress("UnusedParameter")
+fun SubwayApp(
+    appState: SubwayAppState,
+    modifier: Modifier = Modifier,
+) {
+    Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
+        Greeting(
+            name = "hoyahozz",
+            modifier = Modifier.padding(innerPadding),
+        )
+    }
+}
+
+@Composable
+fun Greeting(name: String, modifier: Modifier = Modifier) {
+    Text(
+        text = "Hello $name!",
+        modifier = modifier,
+    )
+}
